feat (ass 9): Create Event and Update Title EndPoints + tests.

+++ IntegrationTests/WebAPI/VeaWebApplicationFactory.cs
@@ -1,5 +1,6 @@
 ﻿using EfcDmPersistence;
 using EfcQueries.Models;
+using EfcQueries.SeedFactories;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
@@ -14,7 +15,7 @@ namespace IntegrationTests.WebAPI;
 
 internal class VeaWebApplicationFactory : WebApplicationFactory<Program>
 {
-    private IServiceCollection serviceCollection;
+    private IServiceCollection? serviceCollection;
 
     protected override void ConfigureWebHost(IWebHostBuilder builder)
     {
@@ -30,8 +31,10 @@ internal class VeaWebApplicationFactory : WebApplicationFactory<Program>
 
             string connString = GetConnectionString();
             services.AddDbContext<DmContext>(options => { options.UseSqlite(connString); });
+            services.AddScoped<DbContext>(sp => sp.GetRequiredService<DmContext>());
             services.AddDbContext<VeaReadModelContext>(options => { options.UseSqlite(connString); });
 
+            services.RemoveAll<ISystemTime>();
             services.AddScoped<ISystemTime, FakeSystemTime>();
 
             SetupCleanDatabase(services);
@@ -40,10 +43,18 @@ internal class VeaWebApplicationFactory : WebApplicationFactory<Program>
     
     private void SetupCleanDatabase(IServiceCollection services)
     {
-        DmContext dmContext = services.BuildServiceProvider().GetService<DmContext>()!;
+        using var provider = services.BuildServiceProvider();
+        using var scope = provider.CreateScope();
+        
+        var dmContext = scope.ServiceProvider.GetRequiredService<DmContext>();
         dmContext.Database.EnsureDeleted();
         dmContext.Database.EnsureCreated();
-        // could seed database here?
+
+        // Seeding DB
+        var veaReadModelContext = scope.ServiceProvider.GetRequiredService<VeaReadModelContext>();
+        veaReadModelContext.Guests.AddRange(GuestSeedFactory.CreateGuests());
+        veaReadModelContext.VeaEvents.AddRange(EventSeedFactory.CreateEvents());
+        veaReadModelContext.SaveChanges();
     }
 
     private string GetConnectionString()
@@ -54,9 +65,13 @@ internal class VeaWebApplicationFactory : WebApplicationFactory<Program>
 
     protected override void Dispose(bool disposing)
     {
-        // clean up the database
-        DmContext dmContext = serviceCollection.BuildServiceProvider().GetService<DmContext>()!;
-        dmContext.Database.EnsureDeleted();
+        if (disposing && serviceCollection is not null)
+        {
+            // clean up the database
+            DmContext dmContext = serviceCollection.BuildServiceProvider().GetService<DmContext>()!;
+            dmContext.Database.EnsureDeleted();
+        }
+        
         base.Dispose(disposing);
     }
